requirements.txt
streamlit
groq
langchain
langchain-core
langsmith
langchain-groq
python-dotenv
sqlite-utils
requests